added in symlink for ripgrep

+++ modules/module.nix
@@ -10,6 +10,10 @@ with lib;
       type = package;
       default = pkgs.nodejs-14_x;
     };
+    ripPath = mkOption {
+      type = package;
+      default = pkgs.ripgrep;
+    };
     findPackage = mkOption {
       type = package;
       default = pkgs.findutils;
@@ -25,22 +29,27 @@ with lib;
         set -euo pipefail
         PATH=${makeBinPath (with pkgs; [ coreutils inotify-tools ])}
         bin_dir=~/.vscode-server/bin
-        [[ -e $bin_dir ]] &&
-        ${findPath} "$bin_dir" -mindepth 2 -maxdepth 2 -name node -type f -exec ln -sfT ${nodePath} {} \; ||
-        mkdir -p "$bin_dir"
-
-        while IFS=: read -r bin_dir event; do
-          # A new version of the VS Code Server is being created.
-          if [[ $event == 'CREATE,ISDIR' ]]; then
-            # Create a trigger to know when their node is being created and replace it for our symlink.
-            touch "$bin_dir/node"
-            inotifywait -qq -e DELETE_SELF "$bin_dir/node"
-            ln -sfT ${nodePath} "$bin_dir/node"
-          # The monitored directory is deleted, e.g. when "Uninstall VS Code Server from Host" has been run.
-          elif [[ $event == DELETE_SELF ]]; then
-            # See the comments above Restart in the service config.
-            exit 0
-          fi
+        bin_dir2=~/.vscode-server-insiders/bin
+        if [[ -e $bin_dir ]]; then
+          ${findPath} "$bin_dir" -mindepth 2 -maxdepth 2 -name node -type f -exec ln -sfT ${nodePath} {} \;
+          ${findPath} "$bin_dir" -path '*/vscode-ripgrep/bin/rg' -exec ln -sfT ${ripPath}/bin/rg {} \;
+          ${findPath} "$bin_dir2" -mindepth 2 -maxdepth 2 -name node -type f -exec ln -sfT ${nodePath} {} \;
+          ${findPath} "$bin_dir2" -path '*/vscode-ripgrep/bin/rg' -exec ln -sfT ${ripPath}/bin/rg {} \;
+        else
+          mkdir -p "$bin_dir"
+          while IFS=: read -r bin_dir event; do
+            # A new version of the VS Code Server is being created.
+            if [[ $event == 'CREATE,ISDIR' ]]; then
+              # Create a trigger to know when their node is being created and replace it for our symlink.
+              touch "$bin_dir/node"
+              inotifywait -qq -e DELETE_SELF "$bin_dir/node"
+              ln -sfT ${nodePath} "$bin_dir/node"
+              ln -sfT ${ripPath} "$bin_dir/node_modules/vscode-ripgrep/bin/rg"
+            # The monitored directory is deleted, e.g. when "Uninstall VS Code Server from Host" has been run.
+            elif [[ $event == DELETE_SELF ]]; then
+              # See the comments above Restart in the service config.
+              exit 0
+            fi
         done < <(inotifywait -q -m -e CREATE,ISDIR -e DELETE_SELF --format '%w%f:%e' "$bin_dir")
       '';
     in
